acpi: operate primarily on xsdt, not rsdt
Not every ACPI platform will have an RSDT, because
not every ACPI platform has memory < 4GiB. Every HLOS
must use the XSDT when one is present.

Change the ACPI shell commands (dumping) and acpi_add_table
to primarily deal with the XSDT, and treat the RSDT as
optional.

On RISC-V, don't build the RSDT (the same should be true for AArch64).

Signed-off-by: Andrei Warkentin <andreiw@mm.st>