efi_console: sync vidconsole apparent size with EFI console mode
When output goes to both serial and vidconsole, or when
choosing mode 0 with vidconsole, EFI output on vidconsole
will be garbled, as line wrapping and scrolling will not work correctly.

- Introduce vidconsole_set_apparent_size, which allows a vidconsole's
  rows/cols to be (temporarily) changed to a smaller value/
- Introduce efi_console_sync_vidconsole(bool), which invokes
  vidconsole_set_apparent_size with the current mode configuration
  when enabling, and resets vidconsole to original size otherwise.
- Modify do_bootefi_exec to perform efi_console_sync_vidconsole before
  and after the program launch.

I thought about keeping the output centered like Tiano, but keeping the
window always at (0, 0) means not having to clear the screen, which is
always good for usability (i.e. not losing output of a failed EFI app).

Signed-off-by: Andrei Warkentin <andreiw@mm.st>